Rename packet pktcapture -> capture.
